Fix Brush and Baseline. Brush needed ref fix (the old style is deprecated)
