+++ .eslintrc
@@ -1,12 +1,16 @@
 {
   "extends": "airbnb",
+  "globals": {
+    "window": true,
+    "document": true
+  },
   "rules": {
-    "no-mixed-operators": [0],
+    "no-mixed-operators": "off",
     "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
   },
   "parserOptions":{
     "ecmaFeatures": {
       "experimentalObjectRestSpread": true
     }
-  }
+  },
 }
