+++ src/js/util.js
@@ -8,6 +8,15 @@
  *  LICENSE file in the root directory of this source tree.
  */
 
-export default function scaleAsString(scale) {
+export function scaleAsString(scale) {
   return `${scale.domain()}-${scale.range()}`;
+}
+
+// http://stackoverflow.com/a/28857255
+export function getElementOffset(element) {
+  const de = document.documentElement;
+  const box = element.getBoundingClientRect();
+  const top = box.top + window.pageYOffset - de.clientTop;
+  const left = box.left + window.pageXOffset - de.clientLeft;
+  return { top, left };
 }
